chunkingandSqs/lambda/requirements.txt
boto3
pandas
openpyxl
python-docx
pdfplumber
